update clr and center dot lines

+++ pkg/sketch/accrew/cloud.go
@@ -16,7 +16,7 @@ func (c Cloud) Dimensions() (int, int) {
 
 func (c Cloud) Draw(context *gg.Context, rand *rand.Rand) {
 	rows := rand.Float64() * 200
-	hue := uint16(rand.Intn(365))
+	hue := rand.Intn(365)
 	growthFactor := rand.Float64() * 50
 	minGrowth := rand.Float64() * 20
 
@@ -27,7 +27,7 @@ func (c Cloud) Draw(context *gg.Context, rand *rand.Rand) {
 	}
 
 	for i := 0.0; i < rows; i++ {
-		r, g, b := clr.HSV{H: hue, S: uint8(i), V: 70}.RGB()
+		r, g, b := clr.HSV{H: hue, S: int(i), V: 70}.RGB()
 		context.SetRGB(float64(r), float64(g), float64(b))
 		for i, height := range accrew {
 			incBy := rand.Float64()*float64(rows)/growthFactor + minGrowth
