+++ pkg/sketch/primitives/line-coloring.go
@@ -18,7 +18,7 @@ func (lc LineColoring) Draw(context *gg.Context, rand *rand.Rand) {
 	rows := rand.Float64() * 200
 	margin := rand.Float64()*200 + 10
 	noiseFactor := rand.Float64() * 10
-	hue := uint16(rand.Intn(365))
+	hue := rand.Intn(365)
 
 	avaliableSpace := 1000 - margin*2
 	spacing := avaliableSpace / rows
@@ -29,7 +29,7 @@ func (lc LineColoring) Draw(context *gg.Context, rand *rand.Rand) {
 	for x := margin; x < 1000-margin; x += spacing {
 		for y := margin; y < 1000-margin; y += spacing {
 			n := (rand.Float64()*2 - 1) * noiseFactor
-			r, g, b := clr.HSV{H: hue, S: uint8(n), V: 70}.RGB()
+			r, g, b := clr.HSV{H: hue, S: int(n), V: 70}.RGB()
 			context.SetRGB(float64(r), float64(g), float64(b))
 			context.DrawLine(x+n, y+n, x-n, y-n)
 			context.Stroke()
