+++ pkg/sketch/primitives/rotated-lines.go
@@ -20,7 +20,7 @@ func (rl RotatedLines) Draw(context *gg.Context, rand *rand.Rand) {
 	rows := math.Floor(rand.Float64() * 200)
 	margin := rand.Float64()*200 + 10
 	noiseFactor := rand.Float64() * 10
-	hue := uint16(rand.Intn(365))
+	hue := rand.Intn(365)
 
 	avaliableSpace := 1000 - margin*2
 	spacing := avaliableSpace / rows
@@ -36,7 +36,7 @@ func (rl RotatedLines) Draw(context *gg.Context, rand *rand.Rand) {
 		for y := margin; y <= 1000-margin; y += spacing {
 			context.Push()
 			n := (rand.Float64()*2 - 1) * noiseFactor
-			r, g, b := clr.HSV{H: hue, S: uint8(n), V: 70}.RGB()
+			r, g, b := clr.HSV{H: hue, S: int(n), V: 70}.RGB()
 			context.SetRGB(float64(r), float64(g), float64(b))
 			context.RotateAbout(n, x, y)
 			context.DrawLine(x+n, y+n, x-n, y-n)
