+++ pkg/sketch/sampling/rectangle.go
@@ -21,8 +21,8 @@ func (c UniformRectangleDot) Dimensions() (int, int) {
 func (c UniformRectangleDot) Draw(context *gg.Context, r *rand.Rand) {
 	rows := 1 + math.Floor(r.Float64()*9)
 	margin := r.Float64() * 0.10
-	startHue := uint16(r.Intn(365))
-	endHue := uint16(r.Intn(365))
+	startHue := r.Intn(365)
+	endHue := r.Intn(365)
 	gaps := r.Float64() * 0.5
 	fmt.Printf("\tGapFactor: %f\n", gaps)
 
@@ -42,7 +42,7 @@ func (c UniformRectangleDot) Draw(context *gg.Context, r *rand.Rand) {
 
 		rect.ShrinkHorizontally(gaps)
 
-		hsv := clr.HSV{H: startHue + uint16(float64(endHue)/(rows-1)*i), S: 60, V: uint8(20 + i*(80.0/(rows-1)))}
+		hsv := clr.HSV{H: startHue + int(float64(endHue)/(rows-1)*i), S: 60, V: int(20 + i*(80.0/(rows-1)))}
 		r, g, b := hsv.RGB()
 		context.SetRGB(float64(r), float64(g), float64(b))
 
@@ -60,7 +60,7 @@ func (c RadialRectangleDot) Dimensions() (int, int) {
 func (c RadialRectangleDot) Draw(context *gg.Context, r *rand.Rand) {
 	rows := 1 + math.Floor(r.Float64()*15)
 	margin := r.Float64() * 0.10
-	hue := uint16(r.Intn(365))
+	hue := r.Intn(365)
 	gaps := r.Float64()
 
 	fmt.Printf("\tGapFactor: %f\n", gaps)
@@ -80,7 +80,7 @@ func (c RadialRectangleDot) Draw(context *gg.Context, r *rand.Rand) {
 		}
 		rect.ShrinkHorizontally(gaps)
 
-		r, g, b := clr.HSV{H: hue, S: uint8(i * 7), V: 70}.RGB()
+		r, g, b := clr.HSV{H: hue, S: int(i * 7), V: 70}.RGB()
 		context.SetRGB(float64(r), float64(g), float64(b))
 
 		filler.Fill(context, rect)
